requirements.txt
google-assistant-grpc==0.0.2
grpc-google-cloud-speech-v1beta1==0.14.0
protobuf==3.1.0
configargparse==0.11.0
phue==0.9
rgbxy==0.5
google-auth-oauthlib==0.1.0
paho-mqtt==1.3.1
ruamel.yaml==0.15.34
rpi-rf==0.9.6
dialogflow==0.1.0
boto3==3-1.4
telepot==12.5
